feat: implement favorite news business logic

+++ app/src/main/java/com/fgr/neonews/data/room/dao/NewsDao.kt
@@ -14,13 +14,13 @@ interface NewsDao {
     suspend fun insertFavorite(news: NewsTable)
 
     @Query("SELECT COUNT(*) FROM tb_news WHERE id = :id")
-    suspend fun isFavorite(id: Int): Boolean
+    suspend fun isFavorite(id: Long): Boolean
 
     @Query("SELECT * FROM tb_news ORDER BY createdAt DESC")
     fun getAllFavorite(): LiveData<List<NewsTable>>
 
     @Query("DELETE FROM tb_news WHERE id = :newsId")
-    suspend fun deleteFavoriteById(newsId: Int)
+    suspend fun deleteFavoriteById(newsId: Long)
 
     @Query("DELETE FROM tb_news")
     suspend fun deleteAllFavorite()
